change code+

+++ src/main/java/com/nimko/shppmentorpracktic6/controller/PersonController.java
@@ -18,11 +18,7 @@ import java.util.List;
 public class PersonController {
 
     @Autowired
-    PersonRepo personRepo;
-
-
-
-
+    private PersonRepo personRepo;
 
     @GetMapping("/{ipn}")
     public Person getPerson(@PathVariable String ipn){
@@ -54,22 +50,21 @@ public class PersonController {
             personRepo.save(person);
             log.info("{}",person);
         } else{
-            return ResponseEntity.status(HttpStatus.UPGRADE_REQUIRED).body(result.toString());
+            return ResponseEntity.badRequest().body(result.toString());
         }
         return ResponseEntity.status(HttpStatus.UPGRADE_REQUIRED).body("Put: " + person);
     }
 
     @DeleteMapping("/{ipn}")
     public ResponseEntity<String> deletePerson(@PathVariable String ipn){
-        Person person;
-        try{
-            log.info("delete {}", ipn);
-            person = personRepo.findPersonByIpn(ipn);
-            personRepo.delete(person);
-        } catch (Exception e){
-            return new ResponseEntity<>(e.toString(),null,HttpStatus.NOT_FOUND);
-        }
-        return new ResponseEntity<>("Delete: " + person, null,HttpStatus.NON_AUTHORITATIVE_INFORMATION);
+        Person person = personRepo.findPersonByIpn(ipn);
+        if(person == null)
+            return ResponseEntity.status(HttpStatus.BAD_GATEWAY)
+                    .body("Person don't exist with IPN: " + ipn);
+        personRepo.delete(person);
+        log.info("delete {}", ipn);
+        return ResponseEntity.status(HttpStatus.NON_AUTHORITATIVE_INFORMATION)
+                .body("Delete: " + person);
     }
 
 }
